requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docker==7.1.0
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.47
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
openai==1.68.2
orjson==3.10.15
packaging==24.2
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.39
tenacity==9.0.0
tqdm==4.67.1
typer==0.15.2
typing-extensions==4.12.2
urllib3==2.3.0
zstandard==0.23.0
langchain-community
langchain-openai
litellm